Refactor validation code
- Add a colour to the rescue case when reading the config
  file fails.
- Since we'll want to add in validation for the global banner,
  update the validator to rely on a shared base class, and
  make a method of the rake output.
