+++ Rakefile
@@ -15,12 +15,9 @@ RSpec::Core::RakeTask.new
 require "govuk_web_banners/validators/recruitment_banner"
 require "rainbow"
 
-desc "show errors in the live config"
-task :check_config do
-  validator = GovukWebBanners::Validators::RecruitmentBanner.new(GovukWebBanners::RecruitmentBanner.all_banners)
-
+def output_validator_info(validator, name)
   if !validator.valid?
-    puts Rainbow("\nLive config contains errors!").red
+    puts Rainbow("\nLive #{name} config contains errors!").red
     validator.errors.each_key do |key|
       puts(key)
       validator.errors[key].each { |error| puts(" - #{error}") }
@@ -28,19 +25,25 @@ task :check_config do
     puts
     exit(1)
   elsif validator.warnings?
-    puts Rainbow("\nLive config is valid, but with warnings").yellow
+    puts Rainbow("\nLive #{name} config is valid, but with warnings").yellow
     validator.warnings.each_key do |key|
       puts(key)
       validator.warnings[key].each { |warnings| puts(" - #{warnings}") }
     end
     puts
   else
-    puts Rainbow("\nLive config is valid!\n").green
+    puts Rainbow("\nLive #{name} config is valid!\n").green
   end
+end
+
+desc "show errors in the live config"
+task :check_recruitment_config do
+  validator = GovukWebBanners::Validators::RecruitmentBanner.new(GovukWebBanners::RecruitmentBanner.all_banners)
+  output_validator_info(validator, "recruitment banner")
 rescue StandardError => e
   puts(e)
-  puts("Live config could not be read (if there are no banners, check banner key is marked as an empty array - banners: [])")
+  puts Rainbow("Live recruitment banner config could not be read (if there are no banners, check banner key is marked as an empty array - banners: [])").red
   exit(1)
 end
 
-task default: %i[check_config rubocop spec]
+task default: %i[check_recruitment_config rubocop spec]
