+++ lib/govuk_web_banners/validators/recruitment_banner.rb
@@ -1,41 +1,12 @@
 require "active_support/core_ext/integer/time"
 require "open-uri"
 
+require "govuk_web_banners/validators/base"
+
 module GovukWebBanners
   module Validators
-    class RecruitmentBanner
-      attr_reader :errors, :warnings
-
-      def initialize(banners)
-        @errors = {}
-        @warnings = {}
-
-        validate(banners)
-      end
-
-      def valid?
-        @errors.keys.none?
-      end
-
-      def warnings?
-        @warnings.keys.any?
-      end
-
+    class RecruitmentBanner < GovukWebBanners::Validators::Base
     private
-
-      def add_error(banner, error)
-        @errors[safe_name(banner)] ||= []
-        @errors[safe_name(banner)] << error
-      end
-
-      def add_warning(banner, warning)
-        @warnings[safe_name(banner)] ||= []
-        @warnings[safe_name(banner)] << warning
-      end
-
-      def safe_name(banner)
-        banner.name || "unnamed banner"
-      end
 
       def validate(banners)
         banners.each do |banner|
@@ -65,11 +36,6 @@ module GovukWebBanners
 
           add_warning(banner, "is expired") unless banner.end_date >= Time.now
         end
-      end
-
-      def overlap?(banner, other_banner)
-        other_banner.start_date < banner.end_date &&
-          other_banner.end_date > banner.start_date
       end
     end
   end
